And V...4, I think?
TODO's include finishing the template for the ini-file, but I am not sure I want them all running either, and haven't decided which sets I will omit and which I will allow to be generated by this. Knittr is complex enough that I can't include _everything_, so I have to draw the line somewhere.